+++ concordium-consensus/src/Concordium/GlobalState/Persistent/Account/StructureV1.hs
@@ -42,6 +42,7 @@ import Concordium.Types.Conditionally
 import Concordium.Types.Execution
 import Concordium.Types.HashableTo
 import Concordium.Types.Parameters
+import Concordium.Types.Tokens
 import Concordium.Utils
 
 import Concordium.GlobalState.Account hiding (addIncomingEncryptedAmount, addToSelfEncryptedAmount, replaceUpTo)