+++ concordium-consensus/src/Concordium/Scheduler.hs
@@ -106,6 +106,7 @@ import Concordium.Scheduler.WasmIntegration.V1 (ReceiveResultData (rrdCurrentSta
 import Concordium.Types.Accounts
 import Concordium.Wasm (IsWasmVersion)
 import qualified Concordium.Wasm as GSWasm
+import Data.Either (isLeft)
 import Data.Proxy
 import Prelude hiding (exp, mod)
 
@@ -384,10 +385,10 @@ dispatchTransactionBody msg senderAccount checkHeaderCost = do
                                 handleConfigureDelegation (mkWTC TTConfigureDelegation) cdCapital cdRestakeEarnings cdDelegationTarget
                         TokenHolder{..} ->
                             onlyWithPLT $
-                                handleTokenHolder (mkWTC TTTokenHolder) thTokenSymbol thOperations
+                                handleTokenHolder (mkWTC TTTokenHolder) thTokenId thOperations
                         TokenGovernance{..} ->
                             onlyWithPLT $
-                                handleTokenGovernance (mkWTC TTTokenGovernance) tgTokenSymbol tgOperations
+                                handleTokenGovernance (mkWTC TTTokenGovernance) tgTokenId tgOperations
   where
     -- Function @onlyWithoutDelegation k@ fails if the protocol version @MPV m@ supports
     -- delegation. Otherwise, it continues with @k@, which may assume the chain parameters version
@@ -2700,34 +2701,38 @@ handleTokenHolder depositContext tokenId tokenOperations =
         let moduleRef = Token._pltModule configuration
         -- TODO Tick energy for loading the module into memory based on the module size. (Issue https://linear.app/concordium/issue/COR-1337)
         -- Invoke the token module with operations.
-        lift $ invokeTokenHolderOperations moduleRef tokenIndex senderAccount tokenOperations
+        (energy, _energyLimitReason) <- getEnergy
+        (res, energyUsed) <- lift $ invokeTokenHolderOperations energy moduleRef tokenIndex senderAccount tokenOperations
+        tickEnergy energyUsed
+        return res
     -- Process the successful transaction computation.
     commitTransaction computeState computeResult = do
         (usedEnergy, energyCost) <- computeExecutionCharge (depositContext ^. wtcEnergyAmount) (computeState ^. energyLeft)
         chargeExecutionCost senderAccount energyCost
         let result = case computeResult of
-                Left encodedRejectReason ->
+                Left PLTEOutOfEnergy -> TxReject OutOfEnergy
+                Left (PLTEFail encodedRejectReason) ->
                     TxReject . TokenHolderTransactionFailed $
                         makeTokenModuleRejectReason tokenId encodedRejectReason
-                Right events -> TxSuccess (TokenModuleEvent . uncurry (TokenEvent tokenId) <$> events)
+                Right events -> TxSuccess events
         return (result, energyCost, usedEnergy)
     -- Call the module of the token with the operations and return the events emitted from the token module.
     invokeTokenHolderOperations ::
+        Energy ->
         TokenModuleRef ->
         Token.TokenIndex ->
         IndexedAccount m ->
         TokenParameter ->
-        m (Either PLTTypes.EncodedTokenRejectReason [(TokenEventType, TokenEventDetails)])
-    invokeTokenHolderOperations _ tokenIndex sender parameter = do
-        result <- withBlockStateRollback $ do
+        m (Either (PLTExecutionError PLTTypes.EncodedTokenRejectReason) [Event], Energy)
+    invokeTokenHolderOperations energy _ tokenIndex sender parameter = do
+        withBlockStateRollback $ do
             let tc =
                     TokenModule.TransactionContext
-                        { tcSender = fst sender,
+                        { tcSender = (fst sender, depositContext ^. wtcSenderAddress),
                           tcSenderAddress = depositContext ^. wtcSenderAddress
                         }
-            runPLT tokenIndex $ TokenModule.executeTokenHolderTransaction tc parameter
-        -- TODO: Generate and handle events: https://linear.app/concordium/issue/COR-705/event-logging
-        return ((\() -> []) <$> result)
+            (res, events, energyUsed) <- runPLTWithEnergy tokenIndex energy $ TokenModule.executeTokenHolderTransaction tc parameter
+            return ((events <$ res, energyUsed), isLeft res)
 
 -- | Handler for a token governance transaction.
 handleTokenGovernance ::
@@ -2762,7 +2767,10 @@ handleTokenGovernance depositContext tokenId tokenOperations =
         let moduleRef = Token._pltModule configuration
         -- TODO Tick energy for loading the module into memory based on the module size. (Issue https://linear.app/concordium/issue/COR-1337)
         -- Invoke the token module with operations.
-        lift $ invokeTokenGovernanceOperations moduleRef tokenIndex senderAccount tokenOperations
+        (energy, _energyLimitReason) <- getEnergy
+        (res, energyUsed) <- lift $ invokeTokenGovernanceOperations energy moduleRef tokenIndex senderAccount tokenOperations
+        tickEnergy energyUsed
+        return res
     -- Process the successful transaction computation.
     commitTransaction computeState computeResult = do
         (usedEnergy, energyCost) <-
@@ -2771,28 +2779,29 @@ handleTokenGovernance depositContext tokenId tokenOperations =
                 (computeState ^. energyLeft)
         chargeExecutionCost senderAccount energyCost
         let result = case computeResult of
-                Left encodedRejectReason ->
+                Left PLTEOutOfEnergy -> TxReject OutOfEnergy
+                Left (PLTEFail encodedRejectReason) ->
                     TxReject . TokenGovernanceTransactionFailed $
                         makeTokenModuleRejectReason tokenId encodedRejectReason
-                Right events -> TxSuccess (TokenModuleEvent . uncurry (TokenEvent tokenId) <$> events)
+                Right events -> TxSuccess events
         return (result, energyCost, usedEnergy)
     -- Call the module of the token with the operations and return the events emitted from the token module.
     invokeTokenGovernanceOperations ::
+        Energy ->
         TokenModuleRef ->
         Token.TokenIndex ->
         IndexedAccount m ->
         TokenParameter ->
-        m (Either PLTTypes.EncodedTokenRejectReason [(TokenEventType, TokenEventDetails)])
-    invokeTokenGovernanceOperations _ tokenIndex sender parameter = do
-        result <- withBlockStateRollback $ do
+        m (Either (PLTExecutionError PLTTypes.EncodedTokenRejectReason) [Event], Energy)
+    invokeTokenGovernanceOperations energy _ tokenIndex sender parameter = do
+        withBlockStateRollback $ do
             let tc =
                     TokenModule.TransactionContext
-                        { tcSender = fst sender,
+                        { tcSender = (fst sender, depositContext ^. wtcSenderAddress),
                           tcSenderAddress = depositContext ^. wtcSenderAddress
                         }
-            runPLT tokenIndex $ TokenModule.executeTokenGovernanceTransaction tc parameter
-        -- TODO: Generate and handle events: https://linear.app/concordium/issue/COR-705/event-logging
-        return ((\() -> []) <$> result)
+            (res, events, energyUsed) <- runPLTWithEnergy tokenIndex energy $ TokenModule.executeTokenGovernanceTransaction tc parameter
+            return ((events <$ res, energyUsed), isLeft res)
 
 -- * Chain updates
 
@@ -2967,25 +2976,25 @@ handleCreatePLT updateHeader payload = runExceptT $ do
         lift (getStateAccount governanceAccountAddress) >>= \case
             Nothing -> throwError $ UnknownAccount governanceAccountAddress
             Just govAcct -> return govAcct
-    let tokenSymbol = payload ^. cpltTokenSymbol
-    maybeExistingToken <- lift $ getTokenIndex tokenSymbol
-    when (isJust maybeExistingToken) $ throwError $ DuplicateTokenId tokenSymbol
+    let tokenId = payload ^. cpltTokenId
+    maybeExistingToken <- lift $ getTokenIndex tokenId
+    when (isJust maybeExistingToken) $ throwError $ DuplicateTokenId tokenId
     createResult <- lift . withBlockStateRollback $ do
         let config =
                 PLTConfiguration
-                    { _pltTokenId = tokenSymbol,
+                    { _pltTokenId = tokenId,
                       _pltModule = payload ^. cpltTokenModule,
                       _pltDecimals = payload ^. cpltDecimals,
                       _pltGovernanceAccountIndex = governanceAccountIndex
                     }
         tokenIx <- createToken config
-        runPLT tokenIx $ TokenModule.initializeToken (payload ^. cpltInitializationParameters)
+        (res, events) <- runPLT tokenIx $ TokenModule.initializeToken (payload ^. cpltInitializationParameters)
+        return ((TokenCreated payload : events) <$ res, isLeft res)
     case createResult of
         Left (e :: TokenModule.InitializeTokenError) -> throwError $ TokenInitializeFailure (show e)
-        Right () -> do
+        Right events -> do
             lift incrementPLTUpdateSequenceNumber
-            -- FIXME: Dummy event. https://linear.app/concordium/issue/COR-705/event-logging
-            return $ TxSuccess [UpdateEnqueued (updateEffectiveTime updateHeader) (CreatePLTUpdatePayload payload)]
+            return $ TxSuccess events
 
 handleUpdateCredentials ::
     (SchedulerMonad m) =>