+++ concordium-consensus/src/Concordium/Scheduler/ProtocolLevelTokens/Kernel.hs
@@ -10,6 +10,7 @@ module Concordium.Scheduler.ProtocolLevelTokens.Kernel (
 import Data.Word
 
 import Concordium.Types
+import Concordium.Types.Tokens
 
 import Concordium.GlobalState.Persistent.BlockState.ProtocolLevelTokens
 
@@ -43,6 +44,9 @@ class (PLTKernelQuery m) => PLTKernelUpdate m where
         -- | Returns 'True' if successful, 'False' if the sender had insufficient balance.
         m Bool
 
+    -- | Log a token module event with the specified type and details.
+    logTokenEvent :: TokenEventType -> TokenEventDetails -> m ()
+
 class (PLTKernelUpdate m) => PLTKernelPrivilegedUpdate m where
     -- | Mint a specified amount and deposit it in the specified account.
     --  The return value indicates if this was successful.
@@ -53,6 +57,13 @@ class (PLTKernelUpdate m) => PLTKernelPrivilegedUpdate m where
     --  The return value indicates if this was successful.
     --  Burning can fail if the amount in the account is less than the specified amount to burn.
     burn :: PLTAccount m -> TokenRawAmount -> m Bool
+
+class PLTKernelChargeEnergy m where
+    -- | Reduce the available energy for the PLT module execution. If the
+    --  available energy is smaller than the given amount, the containing
+    --  transaction will abort and the effects of the transaction will be rolled
+    --  back. The energy is charged in any case (also in case of failure).
+    pltTickEnergy :: Energy -> m ()
 
 class PLTKernelFail e m where
     -- | Abort the current operation by raising an error.