+++ concordium-consensus/tests/globalstate/GlobalStateTests/Account.hs
@@ -9,6 +9,7 @@ import Concordium.GlobalState.Persistent.Account.ProtocolLevelTokens
 import Concordium.GlobalState.Persistent.BlobStore
 import Concordium.GlobalState.Persistent.BlockState.ProtocolLevelTokens
 import Concordium.Types.HashableTo
+import Concordium.Types.Tokens
 import Control.Monad.IO.Class
 import qualified Data.ByteString as BS
 import qualified Data.ByteString.Short as SBS