+++ concordium-consensus/tests/scheduler/Spec.hs
@@ -20,6 +20,7 @@ import qualified SchedulerTests.RejectReasonsRustContract (tests)
 import qualified SchedulerTests.SimpleTransferSpec (tests)
 import qualified SchedulerTests.SimpleTransfersTest (tests)
 import qualified SchedulerTests.StakedAmountLocked (tests)
+import qualified SchedulerTests.TokenCreation (tests)
 import qualified SchedulerTests.TokenHolderTransactions (tests)
 import qualified SchedulerTests.TokenModule (tests)
 import qualified SchedulerTests.TransactionExpirySpec (tests)
@@ -115,4 +116,5 @@ main = hspec $ do
     SchedulerTests.SmartContracts.V1.Caller.tests
     SchedulerTests.KonsensusV1.EpochTransition.tests
     SchedulerTests.TokenModule.tests
+    SchedulerTests.TokenCreation.tests
     SchedulerTests.TokenHolderTransactions.tests